__v1.2.3__ * FIX: error thrown when IMAGICK extension not available * FIX: quality default term * FIX: error thrown while uploading featured image * ADD: updated css. * ADD: Universal view now support text,html-scripts,php,shortcodes * * [Enable it from plugin MAIN Settings] * ADD: cast, crew & collection Taxonomy * ADD: enable/disable developer tools open protection * Many Other Security and Performance updates.

+++ fp_movies.php
@@ -5,7 +5,7 @@
  * Description:         A advanced WordPress plugin to publish movies and TV shows. Join Telegram Channel for all Future Updates and support: <a href="https://t.me/FP_MoviesDB">FP MoviesDB</a>
  * Author:              MSHTeam
  * Author URI:          https://t.me/FP_MoviesDB
- * Version:             1.2.2
+ * Version:             1.2.3
  * Text Domain:         fp-moviesdb
  * Requires PHP:        8.1
  * Requires at least:   6.5
@@ -30,6 +30,8 @@ if (!class_exists('MoviePostGenerator')) {
 
             // if (FP_MOVIES_LOGS) { }
             require_once FP_MOVIES_DIR . 'helper/fp_global_error.php';
+            require_once FP_MOVIES_DIR . 'helper/fp_admin_notices.php';
+
             add_action('wp_enqueue_scripts', [$this, 'fp_enqueue_global_scripts']);
             add_action('admin_enqueue_scripts', [$this, 'fp_enqueue_global_scripts']);
 
@@ -47,6 +49,7 @@ if (!class_exists('MoviePostGenerator')) {
         function setup_admin_hooks()
         {
             add_action('admin_init', array($this, 'setup_admin_init_hooks'));
+            add_action('admin_notices', 'fp_show_transient_notice');
             add_action('admin_enqueue_scripts', 'load_assets');
             add_action('admin_menu', 'add_admin_menus');
             add_action('wp_ajax_check_tmdb_id_exists', 'check_tmdb_exists_currentPosts');
@@ -82,6 +85,9 @@ if (!class_exists('MoviePostGenerator')) {
                 'network' => 'on',
                 'quality' => 'on',
                 'resolution' => 'on',
+                'cast' => 'on',
+                'crew' => 'on',
+                'collection' => 'on',
                 'activeClassicEditor' => 'on',
                 'displayAllSizes' => 'off',
             ]);
@@ -115,12 +121,13 @@ if (!class_exists('MoviePostGenerator')) {
         {
             $ajax_url = admin_url('admin-ajax.php', 'https');
             if (!defined('FP_MOVIES_MODE')) define('FP_MOVIES_MODE', 'prod');
-            if (!defined('FP_MOVIES_VERSION')) define('FP_MOVIES_VERSION', '1.2.2');
+            if (!defined('FP_MOVIES_VERSION')) define('FP_MOVIES_VERSION', '1.2.3');
             if (!defined('FP_MOVIES_WP_REQUIRE')) define('FP_MOVIES_WP_REQUIRE', '6.5');
             if (!defined('FP_MOVIES_PHP_REQUIRE')) define('FP_MOVIES_PHP_REQUIRE', '8.1');
-            if (!defined('FP_MOVIES_FILES')) define('FP_MOVIES_FILES', '1.2.2');
+            if (!defined('FP_MOVIES_FILES')) define('FP_MOVIES_FILES', '1.2.3');
             if (!defined('FP_MOVIES_AUTHOR'))  define('FP_MOVIES_AUTHOR',  'WP_DEBUG');
             if (!defined('FP_MOVIES_NAME'))    define('FP_MOVIES_NAME',    'FP Movies');
+            if (!defined('FP_MOVIES_TEXT_DOMAIN')) define('FP_MOVIES_TEXT_DOMAIN', 'fp-movies-plugin');
             if (!defined('FP_MOVIES_AJAX'))    define('FP_MOVIES_AJAX',    $ajax_url);
             if (!defined('FP_MOVIES_URL'))     define('FP_MOVIES_URL',     plugin_dir_url(__FILE__));
             if (!defined('FP_MOVIES_DIR'))     define('FP_MOVIES_DIR',     plugin_dir_path(__FILE__));
@@ -139,19 +146,34 @@ if (!class_exists('MoviePostGenerator')) {
             if (get_option('mtg_fp_api_key')) define('FP_MOVIES_FP_API_KEY', urlencode(get_option('mtg_fp_api_key')));
             if (!defined('FP_MOVIES_ENCRYPTION_METHOD')) define('FP_MOVIES_ENCRYPTION_METHOD', 'AES-256-CBC');
 
-            if (!defined('FP_MOVIES_ENCRYPTION_KEY')) {
+            if (!defined('FP_MOVIES_ENCRYPTION_KEY') || !defined('FP_MOVIES_ENABLE_DEV_PROTECTION')) {
                 $e_key_settings = get_option('mtg_encryption_settings', []);
-                $e_key = isset($e_key_settings['link_encryption_key']) ? $e_key_settings['link_encryption_key'] : false;
-                if ($e_key !== false) {
-                    define('FP_MOVIES_ENCRYPTION_KEY', $e_key);
-                } else {
-                    require_once FP_MOVIES_DIR . 'helper/fp_get_encryption_key.php';
-                    $e_key = generate_encryption_key();
-                    $e_key_settings['link_encryption_key'] = $e_key;
-                    update_option('mtg_encryption_settings', $e_key_settings);
-                    define('FP_MOVIES_ENCRYPTION_KEY', $e_key);
+                if (!defined('FP_MOVIES_ENCRYPTION_KEY')) {
+                    $e_key = isset($e_key_settings['link_encryption_key']) ? $e_key_settings['link_encryption_key'] : false;
+                    if ($e_key !== false) {
+                        define('FP_MOVIES_ENCRYPTION_KEY', $e_key);
+                    } else {
+                        require_once FP_MOVIES_DIR . 'helper/fp_get_encryption_key.php';
+                        $e_key = generate_encryption_key();
+                        $e_key_settings['link_encryption_key'] = $e_key;
+                        update_option('mtg_encryption_settings', $e_key_settings);
+                        define('FP_MOVIES_ENCRYPTION_KEY', $e_key);
+                    }
+                }
+
+                if (!defined('FP_MOVIES_ENABLE_DEV_PROTECTION')) {
+                    $dev_protection_key = isset($e_key_settings['mtg_enable_dev_protection']) ? $e_key_settings['mtg_enable_dev_protection'] : false;
+
+                    if ($dev_protection_key !== false) {
+                        define('FP_MOVIES_ENABLE_DEV_PROTECTION', $dev_protection_key);
+                    } else {
+                        $e_key_settings['mtg_enable_dev_protection'] = 0;
+                        update_option('mtg_encryption_settings', $e_key_settings);
+                        define('FP_MOVIES_ENABLE_DEV_PROTECTION', 0);
+                    }
                 }
             }
+
 
 
             $logs = get_option('mtg_logs_status', 0);
@@ -230,6 +252,9 @@ if (!class_exists('MoviePostGenerator')) {
                 'network' => 'on',
                 'quality' => 'on',
                 'resolution' => 'on',
+                'cast' => 'on',
+                'crew' => 'on',
+                'collection' => 'on',
                 'activeClassicEditor' => 'on',
                 'displayAllSizes' => 'off',
             );
@@ -334,34 +359,6 @@ if (!class_exists('MoviePostGenerator')) {
             wp_enqueue_script('fp-movies-global', esc_url(FP_MOVIES_URL) . 'js/fp_global_log' . $fp_min_m . '.js', array('jquery'), FP_MOVIES_VERSION, true);
             wp_localize_script('fp-movies-global', 'fpAjax', array('ajaxurl' => FP_MOVIES_AJAX));
         }
-
-        // function add_custom_taxonomies_to_menu()
-        // {
-        //     $taxonomies = ['mtg_audio', 'mtg_year', 'mtg_genre', 'mtg_resolution', 'mtg_quality', 'mtg_network'];
-
-        //     foreach ($taxonomies as $taxonomy) {
-        //         $tax = get_taxonomy($taxonomy);
-
-        //         if ($tax === false) {
-        //             error_log("Taxonomy $taxonomy is not registered.");
-        //             continue;
-        //         }
-
-        //         fp_log_error('TAXONOMY: ' . print_r($tax, true));
-
-        //         if ($tax && is_object($tax) && isset($tax->labels) && isset($tax->labels->name)) {
-        //             add_meta_box(
-        //                 'add-' . esc_attr($taxonomy),
-        //                 esc_html($tax->labels->name),
-        //                 'wp_nav_menu_item_taxonomy_meta_box',
-        //                 'nav-menus',
-        //                 'side',
-        //                 'default',
-        //                 array('taxonomy' => $taxonomy)
-        //             );
-        //         }
-        //     }
-        // }
     }
     new MoviePostGenerator;
 }
